Wifi: Added Diva.Data.SQLite to support improved queries also with SQLite.
However, there is an issue with the new Diva.Data.SQLite.dll: On a fresh install, it does not create all tables in the SQLite database. (As a result, you will see "SQLite error no such table" in the log.) As a work-around, run OpenSim once with the default OpenSim.Data.SQLite.dll. That is, comment out any references to "Diva.Data.SQLite.dll" in the INI file, start OpenSim to let it create the required tables and shut it down again. Then re-activate Diva.Data.SQLite.dll in the configuration file and start OpenSim. This procedure is not necessary if you are upgrading with an already existing SQLite database.