Wifi: Basic infrastructure for localization
Localization can be enabled with the new configuration option "LocalizationCachingPeriod".
HTTP header Accept-Language is used to determine what language should be
presented to the user.
The new configuration option "AdminLanguage" determines what language to use
for messages to the administrator (e.g. in emails).
Localization expects translation data to be present in satellite assemblies.
If there are not any resources available for the specified preferred languages,
or if localization is disabled, then the hard-coded English text is used.