Prevent recursions from crashing stack
This is a bandaid more than a true fix, but without preventing recursions from running wild the process will eventually exceed stack and crash. This was reported as being an issue with loading restricted pages while not logged in to wifi. The actual issue is thus elsewhere in the page handling. Nonetheless recursions should be avoided if they have no bail condition that turns true before hitting stack limits. Please fix the page handling for this and perhaps rethink the approach of using recursions for parsing the pages in the first place.